clean up of config files filter reorder - adds `addsPlugins` - rename `date.js` to `dates.js`
