+++ src/__flightdeck/plugins.js
@@ -1,0 +1,5 @@
+const embedEverything = require("eleventy-plugin-embed-everything");
+
+module.exports = (config) => {
+  config.addPlugin(embedEverything);
+};
